+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Remove dependency on `protolude`
 - Fix parsing bug with Wildcard Selector
+- Implement Descendant Segment
 
 ## 0.1.0.0